moved logout function to the auth api, makes more sense
